# swagger_docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Cabin Booking API",
        "description": "REST API \u0434\u043b\u044f \u0431\u0440\u043e\u043d\u0438\u0440\u043e\u0432\u0430\u043d\u0438\u044f \u0434\u043e\u043c\u0438\u043a\u043e\u0432 (\u043a\u0430\u0431\u0438\u043d)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": ""
        }
    ],
    "paths": {
        "/api/bookings": {
            "get": {
                "operationId": "api_bookings_get_collection",
                "tags": [
                    "Booking"
                ],
                "responses": {
                    "200": {
                        "description": "Booking collection",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "type": "object",
                                    "description": "Booking.jsonld collection.",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/HydraCollectionBaseSchema"
                                        },
                                        {
                                            "type": "object",
                                            "required": [
                                                "member"
                                            ],
                                            "properties": {
                                                "member": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Booking.jsonld"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "summary": "Retrieves the collection of Booking resources.",
                "description": "Retrieves the collection of Booking resources.",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The collection page number",
                        "required": false,
                        "deprecated": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "style": "form",
                        "explode": false
                    }
                ]
            },
            "post": {
                "operationId": "api_bookings_post",
                "tags": [
                    "Booking"
                ],
                "responses": {
                    "201": {
                        "description": "Booking resource created",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking.jsonld"
                                }
                            }
                        },
                        "links": {}
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error.jsonld"
                                }
                            },
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "links": {}
                    },
                    "422": {
                        "description": "An error occurred",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConstraintViolation.jsonld"
                                }
                            },
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConstraintViolation"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConstraintViolation"
                                }
                            }
                        },
                        "links": {}
                    }
                },
                "summary": "Creates a Booking resource.",
                "description": "Creates a Booking resource.",
                "parameters": [],
                "requestBody": {
                    "description": "The new Booking resource",
                    "content": {
                        "application/ld+json": {
                            "schema": {
                                "$ref": "#/components/schemas/Booking"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/api/bookings/{id}": {
            "get": {
                "operationId": "api_bookings_id_get",
                "tags": [
                    "Booking"
                ],
                "responses": {
                    "200": {
                        "description": "Booking resource",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking.jsonld"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error.jsonld"
                                }
                            },
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "links": {}
                    }
                },
                "summary": "Retrieves a Booking resource.",
                "description": "Retrieves a Booking resource.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking identifier",
                        "required": true,
                        "deprecated": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple",
                        "explode": false
                    }
                ]
            },
            "delete": {
                "operationId": "api_bookings_id_delete",
                "tags": [
                    "Booking"
                ],
                "responses": {
                    "204": {
                        "description": "Booking resource deleted"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error.jsonld"
                                }
                            },
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "links": {}
                    }
                },
                "summary": "Removes the Booking resource.",
                "description": "Removes the Booking resource.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking identifier",
                        "required": true,
                        "deprecated": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple",
                        "explode": false
                    }
                ]
            },
            "patch": {
                "operationId": "api_bookings_id_patch",
                "tags": [
                    "Booking"
                ],
                "responses": {
                    "200": {
                        "description": "Booking resource updated",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking.jsonld"
                                }
                            }
                        },
                        "links": {}
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error.jsonld"
                                }
                            },
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "links": {}
                    },
                    "422": {
                        "description": "An error occurred",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConstraintViolation.jsonld"
                                }
                            },
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConstraintViolation"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConstraintViolation"
                                }
                            }
                        },
                        "links": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/ld+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error.jsonld"
                                }
                            },
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "links": {}
                    }
                },
                "summary": "Updates the Booking resource.",
                "description": "Updates the Booking resource.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking identifier",
                        "required": true,
                        "deprecated": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple",
                        "explode": false
                    }
                ],
                "requestBody": {
                    "description": "The updated Booking resource",
                    "content": {
                        "application/merge-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/Booking.jsonMergePatch"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Booking": {
                "type": "object",
                "properties": {
                    "id": {
                        "readOnly": true,
                        "type": "integer"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/User"
                    },
                    "cabin": {
                        "type": "string",
                        "format": "iri-reference",
                        "example": "https://example.com/"
                    },
                    "comment": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Booking.jsonMergePatch": {
                "type": "object",
                "properties": {
                    "id": {
                        "readOnly": true,
                        "type": "integer"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/User"
                    },
                    "cabin": {
                        "type": "string",
                        "format": "iri-reference",
                        "example": "https://example.com/"
                    },
                    "comment": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Booking.jsonld": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/HydraItemBaseSchema"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "id": {
                                "readOnly": true,
                                "type": "integer"
                            }
                        }
                    }
                ],
                "properties": {
                    "owner": {
                        "$ref": "#/components/schemas/User.jsonld"
                    },
                    "cabin": {
                        "type": "string",
                        "format": "iri-reference",
                        "example": "https://example.com/"
                    },
                    "comment": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ConstraintViolation": {
                "type": "object",
                "description": "Unprocessable entity",
                "properties": {
                    "status": {
                        "default": 422,
                        "type": "integer"
                    },
                    "violations": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "propertyPath": {
                                    "type": "string",
                                    "description": "The property path of the violation"
                                },
                                "message": {
                                    "type": "string",
                                    "description": "The message associated with the violation"
                                }
                            }
                        }
                    },
                    "detail": {
                        "readOnly": true,
                        "type": "string"
                    },
                    "type": {
                        "readOnly": true,
                        "type": "string"
                    },
                    "title": {
                        "readOnly": true,
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "instance": {
                        "readOnly": true,
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
            "ConstraintViolation.jsonld": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/HydraItemBaseSchema"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "status": {
                                "default": 422,
                                "type": "integer"
                            },
                            "violations": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "propertyPath": {
                                            "type": "string",
                                            "description": "The property path of the violation"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "The message associated with the violation"
                                        }
                                    }
                                }
                            },
                            "detail": {
                                "readOnly": true,
                                "type": "string"
                            },
                            "description": {
                                "readOnly": true,
                                "type": "string"
                            },
                            "type": {
                                "readOnly": true,
                                "type": "string"
                            },
                            "title": {
                                "readOnly": true,
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "instance": {
                                "readOnly": true,
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        }
                    }
                ],
                "description": "Unprocessable entity"
            },
            "Error": {
                "type": "object",
                "description": "A representation of common errors.",
                "properties": {
                    "title": {
                        "readOnly": true,
                        "description": "A short, human-readable summary of the problem.",
                        "type": "string"
                    },
                    "detail": {
                        "readOnly": true,
                        "description": "A human-readable explanation specific to this occurrence of the problem.",
                        "type": "string"
                    },
                    "status": {
                        "type": "number",
                        "examples": [
                            404
                        ],
                        "default": 400
                    },
                    "instance": {
                        "readOnly": true,
                        "description": "A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.",
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "type": {
                        "readOnly": true,
                        "description": "A URI reference that identifies the problem type",
                        "type": "string"
                    }
                }
            },
            "Error.jsonld": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/HydraItemBaseSchema"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "title": {
                                "readOnly": true,
                                "description": "A short, human-readable summary of the problem.",
                                "type": "string"
                            },
                            "detail": {
                                "readOnly": true,
                                "description": "A human-readable explanation specific to this occurrence of the problem.",
                                "type": "string"
                            },
                            "status": {
                                "type": "number",
                                "examples": [
                                    404
                                ],
                                "default": 400
                            },
                            "instance": {
                                "readOnly": true,
                                "description": "A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.",
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "type": {
                                "readOnly": true,
                                "description": "A URI reference that identifies the problem type",
                                "type": "string"
                            },
                            "description": {
                                "readOnly": true,
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        }
                    }
                ],
                "description": "A representation of common errors."
            },
            "HydraCollectionBaseSchema": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/HydraCollectionBaseSchemaNoPagination"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "view": {
                                "type": "object",
                                "properties": {
                                    "@id": {
                                        "type": "string",
                                        "format": "iri-reference"
                                    },
                                    "@type": {
                                        "type": "string"
                                    },
                                    "first": {
                                        "type": "string",
                                        "format": "iri-reference"
                                    },
                                    "last": {
                                        "type": "string",
                                        "format": "iri-reference"
                                    },
                                    "previous": {
                                        "type": "string",
                                        "format": "iri-reference"
                                    },
                                    "next": {
                                        "type": "string",
                                        "format": "iri-reference"
                                    }
                                },
                                "example": {
                                    "@id": "string",
                                    "type": "string",
                                    "first": "string",
                                    "last": "string",
                                    "previous": "string",
                                    "next": "string"
                                }
                            }
                        }
                    }
                ]
            },
            "HydraCollectionBaseSchemaNoPagination": {
                "type": "object",
                "properties": {
                    "totalItems": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "search": {
                        "type": "object",
                        "properties": {
                            "@type": {
                                "type": "string"
                            },
                            "template": {
                                "type": "string"
                            },
                            "variableRepresentation": {
                                "type": "string"
                            },
                            "mapping": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "@type": {
                                            "type": "string"
                                        },
                                        "variable": {
                                            "type": "string"
                                        },
                                        "property": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "required": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
...
        },
...
        "securitySchemes": {}
    },
    "security": [],
...
}